% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fakebr_data}
\alias{fakebr_data}
\title{Fake.Br dataset.}
\format{Um data frame com 7200 linhas e 3 colunas, onde:
\describe{
  \item{name}{nome do arquivo}
  \item{text}{texto retirado da notícia}
  \item{label}{indica se o conteúdo da notícia é falsa ou verdadeira}
}}
\source{
\url{https://github.com/roneysco/Fake.br-Corpus}
}
\usage{
fakebr_data
}
\description{
Um dataset contendo a junção de 7200 notícias coletadas de websites brasileiros.
Onde 3600 correspondem a notícias verdadeiras e 3600 a falsas.
}
\references{
{Monteiro R.A., Santos R.L.S., Pardo T.A.S., de Almeida T.A., Ruiz E.E.S., Vale O.A. (2018) Contributions to the Study of Fake News in Portuguese: New Corpus and Automatic Detection Results. In: Villavicencio A. et al. (eds) Computational Processing of the Portuguese Language. PROPOR 2018. Lecture Notes in Computer Science, vol 11122. Springer, Cham}
}
\keyword{datasets}
